+++ cellpylib/__init__.py
@@ -7,11 +7,11 @@ CellPyLib is a library for working with Cellular Automata.
 For complete documentation, see: https://cellpylib.org
 """
 
-__version__ = "2.3.1"
+__version__ = "2.4.0"
 
 from .ca_functions import BaseRule, AsynchronousRule, ReversibleRule, binary_rule, init_simple, nks_rule, \
     totalistic_rule, plot_multiple, bits_to_int, int_to_bits, init_random, plot, evolve, until_fixed_point, NKSRule, \
-    BinaryRule, TotalisticRule
+    BinaryRule, TotalisticRule, evolve_block
 
 from .rule_tables import random_rule_table, table_walk_through, table_rule
 
@@ -19,7 +19,7 @@ from .entropy import mutual_information, average_cell_entropy, average_mutual_in
     joint_shannon_entropy
 
 from .ca_functions2d import evolve2d, plot2d, plot2d_slice, plot2d_animate, plot2d_spacetime, init_simple2d, \
-    init_random2d, game_of_life_rule
+    init_random2d, game_of_life_rule, evolve2d_block
 
 from .bien import binary_derivative, cyclic_binary_derivative, ktbien, tbien, bien