+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `evolve_block` function
+- Added `evolve2d_block` function
+- Added block CA demos
 
 ## [2.3.1] - 2021-12-25